users/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.6
exceptiongroup==1.1.2
fastapi==0.101.0
h11==0.14.0
idna==3.4
protobuf==4.23.4
protoletariat==3.2.14
psycopg2-binary==2.9.7
pydantic==2.1.1
pydantic-core==2.4.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
typing-extensions==4.7.1
uvicorn==0.23.2
